+++ Packages/com.def.dialog/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.0.1 - 2022.01.05
+
+**Added**
+
+* DialogProcessor has public `KillDialog()` function.
+
+**Changed**
+
+**Fixed**
+
 ## 1.0.0 - 2021.12.29
 
 **Added**