+++ Packages/com.def.menus/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.0.1 - 2022.01.05
+
+**Added**
+
+**Changed**
+
+* Removed MenuInput's reliance on PlayerInput.
+
+**Fixed**
+
 ## 1.0.0 - 2021.12.25
 
 **Added**